started adding scroll panel

+++ ZoranEngine/Core/Resources/FontResource.cpp
@@ -530,7 +530,7 @@ void FontResource::GenerateFromMSDF(const std::vector<uint32_t>& glyphs)
 
 		Glyph glyph = part.glyph;
 		glyph.UVOffset = uvOffset;
-		glyph.translate /= bmpResolution;
+		glyph.translate /= (float)bmpResolution;
 		//glyph.translate = 0;
 		glyphMap->insert({ glyph.glyph,  glyph });
 
@@ -939,7 +939,7 @@ Glyph GlyphForShape(const Shape& shape, uint32_t uni, int resolution, double adv
 	scale = frame / dims;
 	scale *= 0.70f;
 
-	glyph.translate = Vector2D((float)tr.x, (float)tr.y / 2.0);
+	glyph.translate = Vector2D((float)tr.x, (float)tr.y / 2.0f);
 	glyph.bearing = { (float)-l / (float)resolution, (float)-b / (float)resolution };
 	glyph.size = { (float)dims.x / (float)resolution,(float)dims.y / (float)resolution };
 	glyph.invScale = { 1.0f / (float)scale.x, 1.0f / (float)scale.y };
